Change ServiceInstance completely
ServiceInstance was completely long. It does not make sense to split out the parts of a service type or instance FQDN. There is not connection. They could even be in other domains. DNS-SD doesn't even say that the service name should be the same.

If you are interested in the parts of either the service type or instance, use extract.