add validate#isSubtypeOrInstanceName, but this makes a clear a thought issue in this and other functions
It is stated in subtypeOrInstancePatternStr that a backslash must be escaped. That is not true. A backslash must be escaped if you want a backslash in the name.
A 'single' backslash IS the escape character, and is allowed, but makes no sense if the following character is not a dot or backslash.

Should we enforce 'no meaningless escapes'? Or not?

Also, since it is an escape in the DNS label in DNS, should we not resolve the escape in the other code, where the representation is a JS string?