use chai in 1 test file - nope, this isn't it
Standard cannot cope with this. ESLint forbids it, and rightly so

https://github.com/moll/js-must#asserting-on-property-access

This was the reason we changed to must in the first place. Working around this is a Bad Idea.

https://github.com/prodatakey/dirty-chai is a workaround. But that is silly, because it even seems less supported, and has less downloads, than must.

Another alternative, http://shouldjs.github.io, is also more used than must. Let's try that.

Remember that the main reason for leaving must, is that it seems to be abandoned, and has a bad TS definition.